project1_crawler/app/requirements.txt
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.1
atproto==0.0.51
faktory==1.0.0
